use space as leader
If you're going to re-define the leader, it shouldn't overwrite existing commands that aren't synonymous. `;` will repeat the last search from `f`, `F`, `t`, or `T`, and `,` does the same but in the opposite direction as the original search. These are super helpful commands, and overwriting them for a beginner will only encourage bad habits.
`<space>` is hence a much better candidate for `<leader>`.